+++ Vault/src/utopia/vault/model/immutable/Column.scala
@@ -59,6 +59,8 @@ case class Column(propertyName: String, columnName: String, tableName: String, o
     
     override def alternativeNames = Vector(columnName)
     
+    override def isOptional = allowsNull
+    
     override def toString = s"$columnName $dataType ${ if (isPrimary) "PRIMARY KEY " else ""} ${
         if (usesAutoIncrement) "AUTO_INCREMENT " else ""}"